Move schedule out - reusable
