+++ server/queryHandlers/GetEnvironmentScheduleStatus.js
@@ -4,33 +4,11 @@
 let moment = require('moment');
 let cronService = require('modules/cronService')();
 let config = require('config');
-
-function getEnvironmentDetails(query) {
-  const masterAccountName = config.getUserValue('masterAccountName');
-  
-  var sender = require('modules/sender');
-  var childQuery = {
-    name: 'GetDynamoResource',
-    resource: 'ops/environments',
-    key: query.environmentName,
-    accountName: masterAccountName,
-  };
-
-  return sender.sendQuery({ query: childQuery });
-}
-
-function getSchedule(env, date) {
-  if (env.ScheduleAutomatically) {
-    var schedule = cronService.getActionBySchedule(env.DefaultSchedule, date);
-    return schedule ? schedule : 'ON';
-  }
-
-  return env.ManualScheduleUp ? 'ON' : 'OFF';
-}
+let OpsEnvironment = require('models/OpsEnvironment');
 
 function handler(query) {
-  return getEnvironmentDetails(query).then(environmentDetails => {
-    var schedule = getSchedule(environmentDetails.Value, moment(query.date).toDate());
+  return OpsEnvironment.getByName(query.environmentName).then(opsEnvironment => {
+    var schedule = opsEnvironment.getSchedule(moment(query.date).toDate());
     return { status: schedule };
   });
 };
